+++ src/react-components/src/components/01-atoms/Text/Headings/RNSectionTitle.stories.tsx
@@ -2,10 +2,19 @@ import * as React from "react";
 
 import RNSectionTitle from "./RNSectionTitle";
 import { action } from "@storybook/addon-actions";
+import bem from "../../../../utils/bem";
 
 export default {
   title: "Research Now Section Title",
   component: RNSectionTitle
 };
 
-export const researchNowSectionTitle = () => <RNSectionTitle id="hello-page-title" url="hello"/>;
+
+
+export const researchNowSectionTitle = () => <RNSectionTitle>
+{<a className={`${bem("rn-section-title", [])} rn-section-title`} href={"researchNow-home-url"}>
+    <span id={"research-now-title"}>
+      Research<span className={bem("emphasis", [], "rn-section-title")}>Now</span>
+    </span>
+  </a>}
+   </RNSectionTitle>;