+++ src/react-components/src/components/01-atoms/Text/Headings/RNSectionTitle.tsx
@@ -1,23 +1,11 @@
 /* AT-79 Page Title - Level 1
   This is the ResearchNow Variant, which isn't a semantic heading.*/
 
+// This story serves as documentation and example code for ResearchNow's Section Title.
+// Because ResearchNow uses React-Router's Link, the component itself must be passed in for routing to work properly.
+
 import * as React from "react";
-import bem from "../../../../utils/bem";
-import Heading from "./Heading";
 
-export interface RNSectionTitleProps {
-  id: string;
-  url: string;
-  modifiers?: string[];
-  blockName?: string;
-}
-
-export default function RNSectionTitle(props: RNSectionTitleProps) {
-  const { id, url, modifiers, blockName } = props;
-  const baseClass = "rn-section-title";
-  return <a className={`${bem(baseClass, modifiers, blockName)} ${baseClass}`} href={url}>
-    <span id={id}>
-      Research<span className={bem("emphasis", modifiers, baseClass)}>Now</span>
-    </span>
-  </a>;
+export default function RNSectionTitle(props: React.PropsWithChildren<{}>) {
+  return <>{props.children}</>;
 }