+++ src/react-components/src/components/03-organisms/Headers/RNHeaderWithSearch.stories.tsx
@@ -6,22 +6,25 @@ import { action } from "@storybook/addon-actions";
 import { withKnobs, text } from "@storybook/addon-knobs";
 import UnderlineLink from "../../01-atoms/Links/UnderlineLink";
 
+import bem from "../../../utils/bem";
 
 export default {
-  title: "HeaderWithSearch",
+  title: "Research Now Header With Search",
   component: RNHeaderWithSearch,
 };
 
 export const researchNowSearchHeader = () => <RNHeaderWithSearch
   searchButtonId="searchButtonId"
   searchBarAriaLabel="Search research catalog"
+  sectionTitle={<a className={`${bem("rn-section-title", [], "search-header")} rn-section-title`} href={"researchNow-home-url"}>
+    <span id={"research-now-title"}>
+      Research<span className={bem("emphasis", [], "rn-section-title")}>Now</span>
+    </span>
+  </a>}
   advancedSearchElem={<UnderlineLink url={"advanced-search-link"}>Advanced Search</UnderlineLink>}
   searchBarId="searchBarId"
   dropdownId="dropdownId"
   textFieldAriaLabel="Research Now"
-  // headingContent={<span>Research<span className={"rn-section-title__emphasis"}>Now</span></span>}
-  headingId="researchNow-page-title-id"
-  headingUrl="#research-now-url"
   searchDropdownOptions={["Keyword", "Author", "Subject"]}
   searchSubmitHandler={action("searched")}
   selectChangeHandler={action("changed")}