+++ src/react-components/CHANGELOG.md
@@ -8,12 +8,13 @@ Currently, this repo is in Prerelease.  When it is released, this project will a
 ## PRERELEASE
 ### Added 
 - `RN Header With Search` 
+- a story for `RN Section Title`, which is just a collection of styles
 
 ### Changed
 - `Link` -> `LinkHelper` for clarity
 - Bugfixes in `UnderlineLink` and `IconLink`
 - `Link` missing URL error message
-- Consolidated `Header` and `PageTitle`
+- Consolidated `Heading` and `PageTitle`
 - Added ResearchNow specific `SearchBar` error state 
 - A11Y changes for `SearchBar` and `HeaderWithSearch`
 - Added stories for AT-79, AT-264 and AT-3